+++ core/src/com/zaig100/dg/utils/contain/FlimsyTileC.java
@@ -17,7 +17,7 @@ public class FlimsyTileC extends ObjC {
         return tick_sec;
     }
 
-    public void setTick_sec(float tick_sec) {
-        this.tick_sec = tick_sec;
+    public void setTick_sec(double tick_sec) {
+        this.tick_sec = (float) tick_sec;
     }
 }
